routes now assure protection for private notes and prevent edits/deletes on non-owned content

+++ server/routes/story.router.js
@@ -26,38 +26,52 @@ router.get('/', (req, res) => {
 
 router.get('/inWorld/:id', (req, res) => {
   console.log('GET /api/story/inWorld/id');
-  if (req.isAuthenticated()) {
-    const query = `
+  // if (req.isAuthenticated()) {
+  const query = `
       SELECT *
       FROM "stories"
       WHERE "world_id" = $1
     `;
-    const params = [req.params.id];
-    pool.query(query, params)
-      .then((results) => {
-        res.send(results.rows);
-      })
-      .catch((error) => {
-        res.sendStatus(500);
-        console.log(error);
-      })
-  } else {
-    res.sendStatus(403);
-  }
+  const params = [req.params.id];
+  pool.query(query, params)
+    .then((results) => {
+      res.send(results.rows);
+    })
+    .catch((error) => {
+      res.sendStatus(500);
+      console.log(error);
+    })
+  // } else {
+  //   res.sendStatus(403);
+  // }
 })
 
 router.get('/:id', (req, res) => {
   console.log('GET /api/story/id');
   const query = `
-    SELECT "s"."id", "s"."title", "s"."synopsis", "s"."img_url", "w"."name" as "world", "s"."world_id", "g"."name" as "genre"
+    SELECT "s"."id", "s"."title", "s"."synopsis", "s"."img_url", "w"."name" as "world", "s"."world_id", "g"."name" as "genre",
+    CASE WHEN "u"."id" = $1
+         THEN "s"."private_notes"
+         ELSE NULL
+         END AS "private_notes",
+    CASE WHEN "u"."id" = $1
+         THEN true
+         ELSE false
+         END AS "is_owner"
     FROM "stories" as "s"
     LEFT JOIN "genres" as "g"
     ON "s"."genre_id" = "g"."id"
     JOIN "worlds" as "w"
     ON "s"."world_id" = "w"."id"
-    WHERE "s"."id" = $1
+    JOIN "users" as "u"
+    ON "w"."user_id" = "u"."id"
+    WHERE "s"."id" = $2
   `;
-  const params = [req.params.id];
+  let userId = null;
+  if (req.isAuthenticated()) {
+    userId = req.user.id
+  }
+  const params = [userId, req.params.id];
   pool.query(query, params)
     .then((results) => {
       res.send(results.rows);
@@ -92,8 +106,9 @@ router.post('/', (req, res) => {
 
 router.put('/:id', (req, res) => {
   console.log('PUT /api/story/id');
-  const update = req.body;
-  const query = `
+  if (req.isAuthenticated()) {
+    const update = req.body;
+    const query = `
     UPDATE "stories"
     SET "title" = $1,
         "synopsis" = $2,
@@ -101,34 +116,52 @@ router.put('/:id', (req, res) => {
         "img_url" = $4,
         "private_notes" = $5,
         "world_id" = $6
-    WHERE "id" = $7;
+    WHERE "id" = $7
+    AND EXISTS (SELECT 1 
+                FROM "stories"
+                JOIN "worlds" ON "worlds"."id" = "stories"."world_id"
+                JOIN "users" ON "worlds"."user_id" = "users"."id"
+                WHERE "stories"."id" = $7 AND "users"."id" = $8) ;
   `;
-  const params = [update.title, update.synopsis, update.genre_id, update.img_url, update.private_notes, update.world_id, req.params.id];
-  pool.query(query, params)
-    .then(() => {
-      res.sendStatus(202);
-    })
-    .catch((error) => {
-      res.sendStatus(500);
-      console.log(error);
-    })
+    const params = [update.title, update.synopsis, update.genre_id, update.img_url, update.private_notes, update.world_id, req.params.id, req.user.id];
+    pool.query(query, params)
+      .then((results) => {
+        console.log(results);
+        res.sendStatus(202);
+      })
+      .catch((error) => {
+        res.sendStatus(500);
+        console.log(error);
+      })
+  } else {
+    res.sendStatus(403);
+  }
 })
 
 router.delete('/:id', (req, res) => {
   console.log('DELETE /api/story/:id');
-  const query = `
+  if (req.isAuthenticated()) {
+    const query = `
     DELETE FROM "stories"
-    WHERE "id" = $1;
+    WHERE "id" = $1
+    AND EXISTS (SELECT 1 
+                FROM "stories"
+                JOIN "worlds" ON "worlds"."id" = "stories"."world_id"
+                JOIN "users" ON "worlds"."user_id" = "users"."id"
+                WHERE "stories"."id" = $1 AND "users"."id" = $2) ;
   `;
-  const params = [req.params.id];
-  pool.query(query, params)
-    .then(() => {
-      res.sendStatus(204);
-    })
-    .catch((error) => {
-      res.sendStatus(500);
-      console.log(error);
-    })
+    const params = [req.params.id, req.user.id];
+    pool.query(query, params)
+      .then(() => {
+        res.sendStatus(204);
+      })
+      .catch((error) => {
+        res.sendStatus(500);
+        console.log(error);
+      })
+  } else {
+    res.sendStatus(403);
+  }
 })
 
 module.exports = router;
